requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
cycler==0.10.0
discord.py==1.7.3
environs==9.3.2
idna==3.2
kiwisolver==1.3.1
marshmallow==3.12.1
matplotlib==3.4.2
multidict==5.1.0
numpy==1.21.0
pandas==1.2.5
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
typing-extensions==3.10.0.0
yarl==1.6.3
